requirements.txt
fastapi
uvicorn
google-genai